requirements.txt
streamlit
requests
python-dotenv
jsonschema

